added recipe model and test

+++ app/core/models.py
@@ -47,8 +47,8 @@ class Recipe(models.Model):
     )
     title = models.CharField(max_length=255)
     description = models.TextField(blank=True)
-    price = models.IntegerField()
-    time_minutes = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.DecimalField(max_digits=5, decimal_places=2)
+    time_minutes = models.IntegerField()
     link = models.CharField(max_length=255, blank=True)
 
     def __str__(self):
